Rework levels and order
Order is now specified by the actual ordering passed to levels!().
The order of levels in the 'index' field (integer index) is now an
internal property which isn't exposed. levels!() now preserves that
index when adding levels, so that no recoding of existing data or
CategoricalValue objects is needed. Deleting levels still invalidates
all references.