Sort levels on construction from array when possible
We previously preserved their order of appearance, but this is not likely
to be the intended sorting and it can be quite unstable. One caveat is that
e.g. Array{Any} inputs will not be sorted as there is no isless() method for
this element type. Levels added later via setindex!() are still added at the
end of the pool.

copy() needs to be adapted since it previously used the generic AbstractArray
method, which did not preserve levels. The new version also preserves #undef
entries, just like Array.