Merge NominalArray and OrdinalArray into CategoricalArray
This required introducing more constructors instead of relying on the
fallback to convert(): convert() does not copy its input when possible,
but constructors have to make a copy to avoid mutating the input when
calling ordered!(). Also clean reorganize the code base, and improve tests.

Thanks to Jacob Quinn for the first half of this rework.