Fix mergelevels() in case of ambiguous orderings
The fact that all original orderings are preserved does not mean we can
consider merged levels as ordered: ambiguities can remain if some subsequent
pairs of elements do not appear in original levels.

Do not change merge algorithm depending on whether a common order can be found
or not: always proceed the same, and check orderedness after the fact.

More efficient algorithms can likely be found, but for now this is
good enough.