Remove useless in() methods
These are not actually needed for standard collections, which rely on isequal() or == and hash().
They create ambiguities and can even be incorrect, e.g. for ObjectIdDict. They could even be less
efficient than the default if we stored hashes and implemented optimized hash(::CatValue, ::UInt)
methods. For reference, they have been introduced in 9638d31.