Fix similar() to respect requested element type, implement float() and complex() (#133)
similar(::CategoricalArray, T) used to return a CategoricalArray{T}, which is not a correct
implementation of the AbstractArray interface. Change it to return Array{T}, except when
T <: Union{CatValue,Missing}. This also fixes the behavior of AbstractArray constructors and
convert methods, whose fallback definitions use similar().

With these fixes, implementing float() and complex() is trivial.

Stop testing AbstractArray{T}(::CategoricalArray), which is not supported in Base either.