Fix promotion rules for CatValue (#159)
These rules weren't really used previously since concatenation of arrays was used instead,
but packages can now use Tables.allocatecolumns to get a CategoricalArray from a CatValue,
and use promotion rules to get the right eltype from various input values.

It's not clear what some promotion rules involving a CatValue with the reftype unspecified should return. In the absence of a rule, the fallback appears to drop all type parameters, and return Any if Missing is involved. Anyway these should be very rare in practice, so what matters is that they do not fail and return a general enough type to avoid conversion errors.