Remove CategoricalString
Deprecate it in favor of `CategoricalValue{String}`. Keep all deprecations
and add new ones so that most functionality keeps working. The main breaking
change which cannot be avoided is the inheritance from `AbstractString`.

Stop overloading `Base.tostr_sizehint` as it appears to trigger lots of recompilation
and it's just a small optimization (see JuliaLang/julia#33757).