Transform SubString and AbstractString leveltypes to String by default (#264)
Since CategoricalArrays are intended for cases where the number of levels is small,
it makes sense to copy `SubString`s for simplicity. Calling
`CategoricalArray{SubString{String}}` will still allow constructing that array type.
Also replace `AbstractString` with `String` as it is the type of array one gets by
mixing `String` and `SubString` objects.