Store equality and subset comparisons with one pool in a field
When comparing `CategoricalValue`s from different pools, checking whether pools
are equal or whether one is a superset of the other is too costly to be done
repeatedly for all values in an array. To fix this, store the result of the
comparison in new fields, so that subsequent checks are fast. This means that
the check only needs to be done the first time two pools are compared,
or the first time after a comparison with another pool was made (which is rare).
For equality checks, this relies on checking that none of the two pools was modified
since they were marked as equal. For subset checks, this relies on the fact that
pools can only be added new levels, never lose levels.
This approach is simpler and more thread-safe than keeping a global table.

This makes it possible to support relatively efficient comparisons between
`CategoricalValue`s from different pools using `==` and `<`.
Also implement `==`/`isequal` and `hash` on `CategoricalPool`, with `==` being
equivalent to `isequal` (like for `Dict`) for performance and ignoring the
orderedness of the pool (like `CategoricalValue`).