Add `CategoricalValue(x, source)` and disallow mixed `isless` and `<`
`isless` and `<` between `CategoricalValue` and other types can breaks
transitivity. Require converting the value to a `CategoricalValue`,
attaching it to a pool, using
`CategoricalValue(x, source::Union{CategoricalValue, CategoricalArray)`.
Deprecate `CategoricalValue(i::Integer, pool::CategoricalPool)` in favor
of `pool[i]` or `CategoricalValue(pool, i)`, as the order is more logical
and it avoids a potential ambiguity if `source::CategoricalPool` is allowed
in the future.