Allow adding new levels to ordered arrays (#342)
Instead, mark arrays as unordered if one of the pools is unordered,
if pools have incompatible orderings or if orders of all pairs of levels
cannot be determined.
This ensures that operations supported by any `AbstractArray` never fail for
`CategoricalArray`. An additional `protected` setting could be added to
manually decide to lock levels of ordered or unordered `CategoricalArray`s.