Do not print type information when `:compact=>true`
`:compact=>true` used to be recommended only to print fewer digits for numbers,
but now the convention appears to be to use it more generally. This is not
needed for `CatetoricalArray` printing as they already set `:typeinfo`,
but it can help in other contexts such as when `NamedArray` dimension names
are `CategoricalValue`s.