Support reading from and writing to Arrow files
This requires overriding `Arrow.DictEncoding` so that an `Arrow.DictEncoded`
with a `CategoricalArray` dictionary with one entry per level is created.
This is the only way to ensure that indexing the Arrow column gives
`CategoricalValue` objects. In practice such columns will most often
be used after conversion to `CategoricalArray` via `copy`, `DataFrame`, etc.

Apparently, pandas do not allow reading the resulting file if the array
allows for missing values as it does not accept `missing` in the dictionary.
Instead it would need missing entries to be coded via null indices, which
is less efficient.