Choose different quantile cutpoints in `cut(x, n)` (#416)
`Statistics.quantile` returns values which are not the most appropriate
to generate labels. It is more intuitive to choose values from the actual data,
which are likely to have fewer decimals and make more sense for users.
Since intervals are closed on the left, we just have to use the value right
below the quantile. This doesn't change group assignments (only labels).